refactor into separate preview statuses controller
