+++ app/controllers/storages_controller.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
 class StoragesController < ApplicationController
-  before_action :flash_preview_status, only: :index
-  
   def index
     @storages = Storage.all.order(:csv_type)
   end
