+++ app/helpers/application_helper.rb
@@ -66,27 +66,4 @@ module ApplicationHelper
     tags.gsub(/<script /, '<script nonce="**CSP_NONCE**" ').html_safe
   end
   # rubocop:enable Rails/OutputSafety
-
-  def preview_generation_started?
-    PreviewGenerationStatusInformation.exists?
-  end
-
-  def preview_generation_completed?
-    return unless preview_generation_started?
-
-    completed = false
-
-    pgsi = PreviewGenerationStatusInformation.last
-    if pgsi.current_progress.start_with?(PUBLISH_COMPLETE_TEXT) ||
-       pgsi.current_progress.start_with?('There was an error')
-      completed = true
-      PreviewGenerationStatusInformation.delete_all
-      # maintain the indexes and tables in the local, dev & staging envs.
-      # The production env times out and periodic maintenance should be run
-      # in production anyway.
-      PerformInsitutionTablesMaintenanceJob.perform_later unless production?
-    end
-
-    completed
-  end
 end
